# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cylestio Local Server API
  description: |
    API for the Cylestio Local Server, providing telemetry, metrics, and agent management capabilities.
    This API enables developers to integrate with and monitor AI agent operations.
  version: 1.0.0
  contact:
    name: Cylestio Support
    url: https://github.com/cylestio/cylestio-local-server
servers:
  - url: 'http://localhost:8000'
    description: Local Development Server
...
paths:
...
  /v1/telemetry:
    post:
      summary: Submit Telemetry Event
      description: Submits a single telemetry event to the server
      operationId: submitTelemetryEvent
      tags:
        - Telemetry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryEvent'
      responses:
        '201':
          description: Event successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  event_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
    
    get:
      summary: Get Telemetry Events
      description: Retrieves telemetry events matching the specified filters
      operationId: getTelemetryEvents
      tags:
        - Telemetry
      parameters:
        - name: agent_id
          in: query
          description: Filter by agent ID
          required: false
          schema:
            type: string
        - name: from_time
          in: query
          description: Start time for event filtering (ISO 8601 format)
          required: false
          schema:
            type: string
            format: date-time
        - name: to_time
          in: query
          description: End time for event filtering (ISO 8601 format)
          required: false
          schema:
            type: string
            format: date-time
        - name: event_type
          in: query
          description: Filter by event type
          required: false
          schema:
            type: string
        - name: level
          in: query
          description: Filter by event level
          required: false
          schema:
            type: string
            enum: [debug, info, warning, error, critical]
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelemetryEvent'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /v1/telemetry/batch:
    post:
      summary: Submit Batch Telemetry Events
      description: Submits multiple telemetry events in a batch
      operationId: submitBatchTelemetryEvents
      tags:
        - Telemetry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/TelemetryEvent'
              required:
                - events
      responses:
        '201':
          description: Events successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  total:
                    type: integer
                    description: Total number of events in the batch
                    example: 5
                  processed:
                    type: integer
                    description: Number of events successfully processed
                    example: 5
                  failed_events:
                    type: array
                    description: List of events that failed processing
                    items:
                      type: object
                      properties:
                        event:
                          $ref: '#/components/schemas/TelemetryEvent'
                        error:
                          type: string
                          example: "Invalid event format"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /v1/telemetry/events/{event_id}:
    get:
      summary: Get Telemetry Event by ID
      description: Retrieves a specific telemetry event by its ID
      operationId: getTelemetryEventById
      tags:
        - Telemetry
      parameters:
        - name: event_id
          in: path
          description: ID of the telemetry event
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryEvent'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/telemetry/events/by-ids:
    post:
      summary: Get Multiple Telemetry Events by IDs
      description: Retrieves multiple telemetry events by providing a list of event IDs
      operationId: getTelemetryEventsByIds
      tags:
        - Telemetry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_ids:
                  type: array
                  description: List of event IDs to retrieve
                  items:
                    type: string
                    format: uuid
              required:
                - event_ids
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryEvent'
        '400':
          description: Bad request (too many IDs)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  schemas:
    TelemetryEvent:
      type: object
      properties:
        schema_version:
          type: string
          example: "1.0"
        timestamp:
          type: string
          format: date-time
          example: "2023-05-01T12:34:56.789Z"
        trace_id:
          type: string
          example: "abcdef123456"
        span_id:
          type: string
          example: "span123"
        parent_span_id:
          type: string
          example: "parentspan456"
        name:
          type: string
          example: "llm.request"
        level:
          type: string
          enum: [debug, info, warning, error, critical]
          example: "info"
        agent_id:
          type: string
          example: "agent123"
        attributes:
          type: object
          additionalProperties: true
          example:
            model: "gpt-4"
            prompt_tokens: 150
            completion_tokens: 50
      required:
        - schema_version
        - timestamp
        - name
        - level
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
          example: 100
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20
        pages:
          type: integer
          description: Total number of pages
          example: 5
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        status_code:
          type: integer
          description: HTTP status code
          example: 404
        detail:
          type: string
          description: Detailed error information
          example: "The requested resource could not be found"
    ValidationError:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Validation error"
        status_code:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: array
          description: Validation error details
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                example: ["body", "name"]
              msg:
                type: string
                example: "field required"
              type:
                type: string
                example: "value_error.missing"
...